requirements.txt
# torch
# torchaudio
numpy
pandas
jupyter
notebook
tqdm
scikit-learn
opencv-python
tensorboard
kaggle
hydra-core
hydra_colorlog
albumentations
matplotlib
